small fix files and push function in module
